Render user_guide_link from app model in app form header, and move app form header to capture html apps
